fix: timestamp (#20)
Description
---
Timestamp was changed to simple number of seconds.

How Has This Been Tested?
---
Manually by running the tari-explorer.

What process can a PR reviewer use to test or verify this change?
---
Just run it and it should work fine.


<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] Other - Please specify
If you run this on older version that returns timestamp as a structure,
it will not work.
<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->