feat: search commitment/kernels (#23)
Description
---
Added search for kernels.
The search commitment was broken (wrong argument name).
Reworked the errors, all returns json when json is requested.

Motivation and Context
---
tari-explorer is a backend to the new tari explorer, so we need to keep
this up-to-date.

How Has This Been Tested?
---
Manually.

What process can a PR reviewer use to test or verify this change?
---
Run this tari-explorer and test the search, also you can test the
`&json` in the url.

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->