fix: Use default address 'localhost' in Client
The default GRPC address for the base node client was incorrectly set to "127.0.0.1", but "localhost" is correct (and used by the docker-compose setup). This commit changes the default address to "localhost". Also, the client constructor should only use the environment variable if it is defined and should fall back to the default if it is not set.